fix: move `add_events` to be under `Widget` to follow GTK
